requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy[asyncio]==2.0.32
asyncpg==0.29.0
aiosqlite==0.20.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.8.2
pydantic-settings==2.4.0
